* master: (86 commits)
  [ahrs] update/clean test_ahrs
  [python] fix settings path
  [messages] add some descriptions again
  fix name of git submodule
  [chibios] switch back to a repo that have to 2.6 stable
  [chibios] switch back to correct version
  [ocaml] filter for 'field' tag when looking at messages
  [submodules] updated to a new chibios repository
  [messages] remove description node again for now
  [modules] mission_rotorcraft: only check for ned_initialized_i
  [modules][messages] mission: LLA waypoints in 1e7deg instead of float
  [build] don't compile natnet if not needed
  [messages] possibility to add descriptions
  [modules] rename flight_time to time_countdown
  [sim] remove launchsitl symlink
  [rotorcraft] navigation: fix NavCircleCount
  [modules] rotorcraft_cam: fix DEFAULT_MODE
  [omap] start basic spi driver
  [rotorcraft] arming: only check if motors can be armed if ahrs_is_aligned
  [rotorcraft] force MODE_STARTUP instead of KILL until ahrs is aligned
  ...

+++ sw/airborne/firmwares/rotorcraft/navigation.h
@@ -60,7 +60,7 @@ extern int32_t nav_heading; ///< with #INT32_ANGLE_FRAC
 extern float nav_radius;
 
 extern int32_t nav_leg_progress;
-extern int32_t nav_leg_length;
+extern uint32_t nav_leg_length;
 
 extern uint8_t vertical_mode;
 extern uint32_t nav_throttle;  ///< direct throttle from 0:MAX_PPRZ, used in VERTICAL_MODE_MANUAL
@@ -135,7 +135,7 @@ extern void nav_circle(struct EnuCoor_i * wp_center, int32_t radius);
   nav_circle(&waypoints[_center], POS_BFP_OF_REAL(_radius)); \
 }
 
-#define NavCircleCount() (abs(nav_circle_radians) / INT32_ANGLE_2_PI)
+#define NavCircleCount() ((float)abs(nav_circle_radians) / INT32_ANGLE_2_PI)
 #define NavCircleQdr() ({ int32_t qdr = INT32_DEG_OF_RAD(INT32_ANGLE_2_PI_2 - nav_circle_qdr) >> INT32_ANGLE_FRAC; NormCourse(qdr); qdr; })
 
 /** True if x (in degrees) is close to the current QDR (less than 10 degrees)*/
